refactor: merged part of  the search component tests into one file
